Restructure files and add missing declarations
