+++ types/materialize-css/inputfields.d.ts
@@ -1,0 +1,7 @@
+/// <reference path="./common.d.ts" />
+
+declare namespace M {
+    function updateTextFields(): void;
+
+    function textareaAutoResize(textarea: Element): void;
+}
