+++ types/materialize-css/parallax.d.ts
@@ -1,0 +1,14 @@
+/// <reference path="./common.d.ts" />
+
+declare namespace M {
+    class Parallax extends Component<ParallaxOptions> {
+    }
+
+    interface ParallaxOptions {
+        /**
+         * The minimum width of the screen, in pixels, where the parallax functionality starts working
+         * @default 0
+         */
+        responsiveThreshold: number;
+    }
+}
