+++ types/materialize-css/range.d.ts
@@ -1,0 +1,6 @@
+/// <reference path="./common.d.ts" />
+
+declare namespace M {
+    class Range extends Component<undefined> {
+    }
+}
