+++ types/materialize-css/test/materialize-css-global.test.ts
@@ -1,4 +1,5 @@
 const elem = document.querySelector('.whatever')!;
+
 // $ExpectType Sidenav
 const sidenav = new M.Sidenav(elem);
 
@@ -15,7 +16,7 @@ const autocomplete = new M.Autocomplete(elem);
 const characterCounter = new M.CharacterCounter(elem);
 
 // $ExpectType Tooltip
-const tooltips = new M.Tooltip(elem);
+const tooltip = new M.Tooltip(elem);
 
 // $ExpectType FloatingActionButton
 const fab = new M.FloatingActionButton(elem);
@@ -23,14 +24,44 @@ const fab = new M.FloatingActionButton(elem);
 // $ExpectType Toast
 const toast = M.toast({ html: 'I am a toast!' });
 
-// $ExpectType DatePicker
-const datePicker = new M.DatePicker(elem);
+// $ExpectType Datepicker
+const datePicker = new M.Datepicker(elem);
 
-// $ExpectType TimePicker
-const timePicker = new M.TimePicker(elem);
+// $ExpectType Timepicker
+const timePicker = new M.Timepicker(elem);
 
 // $ExpectType Dropdown
 const dropdown = new M.Dropdown(elem);
 
 // $ExpectType FormSelect
-const formSelect = new M.FormSelect(elem);
+const formselect = new M.FormSelect(elem);
+
+// $ExpectType Carousel
+const carousel = new M.Carousel(elem);
+
+// $ExpectType Chips
+const chips = new M.Chips(elem);
+
+// $ExpectType Collapsible
+const collapsible = new M.Collapsible(elem);
+
+// $ExpectType Materialbox
+const materialbox = new M.Materialbox(elem);
+
+// $ExpectType Parallax
+const parallax = new M.Parallax(elem);
+
+// $ExpectType Pushpin
+const pushpin = new M.Pushpin(elem);
+
+// $ExpectType Range
+const range = new M.Range(elem);
+
+// $ExpectType ScrollSpy
+const scrollspy = new M.ScrollSpy(elem);
+
+// $ExpectType Slider
+const slider = new M.Slider(elem);
+
+// $ExpectType TapTarget
+const taptarget = new M.TapTarget(elem);
