+++ types/materialize-css/test/materialize-css-jquery.test.ts
@@ -59,3 +59,27 @@ $(".whatever").dropdown("open");
 $(".whatever").dropdown("close");
 $(".whatever").dropdown("destroy");
 $(".whatever").dropdown("recalculateDimensions");
+
+$(".whatever").carousel();
+
+$(".whatever").dropdown();
+
+$(".whatever").dropdown();
+
+$(".whatever").dropdown();
+
+$(".whatever").dropdown();
+
+$(".whatever").dropdown();
+
+$(".whatever").dropdown();
+
+$(".whatever").dropdown();
+
+$(".whatever").dropdown();
+
+$(".whatever").dropdown();
+
+$(".whatever").dropdown();
+
+$(".whatever").dropdown();
